refactor(io): move secret-flag resolution onto SourceResolver
The stdin-or-file-only rule now has a second caller (runtime, PR 2035), so
resolveSecretFlag moves off the credentials shared module and onto
SourceResolver as resolveSecret. It throws SourceResolutionError (already a
subtype of InputValidationError), which keeps the io package from depending on
handler-level error types.

Also bump the generated CDK app template's aws-cdk-lib to ~2.266.0 so it
satisfies the @aws/agentcore-cdk peer floor once the credential constructs
release.
