+++ src/assets/cdk/package.json
@@ -24,7 +24,7 @@
   },
   "dependencies": {
     "@aws/agentcore-cdk": "0.1.0-alpha.45",
-    "aws-cdk-lib": "~2.261.0",
+    "aws-cdk-lib": "~2.266.0",
     "constructs": "~10.7.0"
   }
 }
